--- blog-api/src/main/java/com/ywoosang/tech/controller/HomeController.java
@@ -1,20 +1,0 @@
-package com.ywoosang.tech.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class HomeController {
-
-    // test
-    @GetMapping("/")
-    public String root() {
-        return "hello root";
-    }
-
-    // test
-    @GetMapping("/home")
-    public String home() {
-        return "hello home";
-    }
-}
